requirements.txt
# Requirements for jaid-agentic-newsletter
openai
pandas
python-dotenv
PyPDF2
Flask